# Security cache dynamic collection for custom properties