… Common/QuantConnect.csproj
… Common/Securities/SecurityCache.cs
… Engine/QuantConnect.Lean.Engine.csproj
… Indicators/QuantConnect.Indicators.csproj
… Report/QuantConnect.Report.csproj
… Research/QuantConnect.Research.csproj
… Tests/Common/Securities/SecurityTests.cs
… Tests/QuantConnect.Tests.csproj
… ToolBox/QuantConnect.ToolBox.csproj