<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->

This adds a custom properties collection carried by `SecurityCache` and helper methods and mechanism to make it accessible from `Security` instances:
- Using `Security.Get<T>/Try<T>Get/Set` methods or the indexer (`Security[]`), the custom properties can be set/accessed by key.
    - `Get/TryGet/Set` allow to set and get (casting to the desired/expected type) the dynamic properties. C# users might use these methods and -- if performance is not that big of an issue -- they could use the dynamic object capabilities (see next point).
- Using `dynamic` the properties can be set/accessed dynamically by property name (e.g. `security.SlowEma`) instead of by key. C# users can rely un this for easy dynamic properties access. Python users can also use this since it will be no different than setting custom Python attributes to a given instance.

`BasicTemplateBenchmark` benchmark algorithm was ran using both QuantConnect/pythonnet version 2.0.18 and 2.0.19. No real difference was found:

Master:

```
Algorithm Id:(BasicTemplateBenchmark) completed in 34.06 seconds at 106k data points per second. Processing total of 3,617,396 data points.

Algorithm Id:(BasicTemplateBenchmark) completed in 34.00 seconds at 106k data points per second. Processing total of 3,617,396 data points.

Algorithm Id:(BasicTemplateBenchmark) completed in 34.05 seconds at 106k data points per second. Processing total of 3,617,396 data points.
```

PR:

```
Algorithm Id:(BasicTemplateBenchmark) completed in 33.44 seconds at 108k data points per second. Processing total of 3,617,396 data points.

Algorithm Id:(BasicTemplateBenchmark) completed in 33.99 seconds at 106k data points per second. Processing total of 3,617,396 data points.

Algorithm Id:(BasicTemplateBenchmark) completed in 33.69 seconds at 107k data points per second. Processing total of 3,617,396 data points.
```

The algorithm was also modified to access several properties of the security objects in `OnData`. Again, no meaningful difference was observed.

```python
class BasicTemplateBenchmark(QCAlgorithm):

    def Initialize(self):
        self.SetStartDate(2000, 1, 1)
        self.SetEndDate(2022, 1, 1)
        self.SetBenchmark(lambda x: 1)
        self.AddEquity("SPY")

    def OnData(self, data):
        # Accessing security properties
        spy = self.Securities["SPY"]
        _open = spy.Open
        close = spy.Close
        high = spy.High
        low = spy.Low
        volume = spy.Volume
        symbol = spy.Symbol
        quote_currency = spy.QuoteCurrency
        symbol_properties = spy.SymbolProperties
        exchange = spy.Exchange
        _type = spy.Type

        if not spy.Holdings.Invested:
            self.SetHoldings("SPY", 1)
            self.Debug("Purchased Stock")
```

Master:
```
Algorithm Id:(BasicTemplateBenchmark) completed in 41.89 seconds at 86k data points per second. Processing total of 3,617,396 data points.

Algorithm Id:(BasicTemplateBenchmark) completed in 40.59 seconds at 89k data points per second. Processing total of 3,617,396 data points.

Algorithm Id:(BasicTemplateBenchmark) completed in 40.99 seconds at 88k data points per second. Processing total of 3,617,396 data points.
```

PR:

```
Algorithm Id:(BasicTemplateBenchmark) completed in 38.90 seconds at 93k data points per second. Processing total of 3,617,396 data points.

Algorithm Id:(BasicTemplateBenchmark) completed in 40.55 seconds at 89k data points per second. Processing total of 3,617,396 data points.

Algorithm Id:(BasicTemplateBenchmark) completed in 39.53 seconds at 92k data points per second. Processing total of 3,617,396 data points.
```

#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->

Closes #6809 

#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->

Reducing some of the boiler plate scaffolding required building multi asset strategies.

#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->

#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->

Unit tests and regression algorithms.

#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Refactor (non-breaking change which improves implementation)
- [ ] Performance (non-breaking change which improves performance. Please add associated performance test and results)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [x] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->